// Qanat.Web/src/app/shared/generated/model/water-account-water-account-contact-update-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class WaterAccountWaterAccountContactUpdateDto { 
    WaterAccountContactID?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface WaterAccountWaterAccountContactUpdateDtoForm { 
    WaterAccountContactID?: FormControl<number>;
}

export class WaterAccountWaterAccountContactUpdateDtoFormControls { 
    public static WaterAccountContactID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}