// Qanat.Web/src/app/shared/generated/model/mapbox-response-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConfidenceEnum } from './confidence-enum';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class MapboxResponseDto { 
    Address?: string | null;
    SecondaryAddress?: string | null;
    City?: string | null;
    State?: string | null;
    StatePostalCode?: string | null;
    ZipCode?: string | null;
    Confidence?: ConfidenceEnum;
    ResponseJsonString?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}
export namespace MapboxResponseDto {
}


export interface MapboxResponseDtoForm { 
    Address?: FormControl<string>;
    SecondaryAddress?: FormControl<string>;
    City?: FormControl<string>;
    State?: FormControl<string>;
    StatePostalCode?: FormControl<string>;
    ZipCode?: FormControl<string>;
    Confidence?: FormControl<ConfidenceEnum>;
    ResponseJsonString?: FormControl<string>;
}
export namespace MapboxResponseDto {
}


export class MapboxResponseDtoFormControls { 
    public static Address = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SecondaryAddress = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static City = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static State = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static StatePostalCode = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ZipCode = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Confidence = (value: FormControlState<ConfidenceEnum> | ConfidenceEnum = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<ConfidenceEnum>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ResponseJsonString = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}
export namespace MapboxResponseDto {
}